Fix Tinycon cross-domain issues and initial setup.
Because the image is coming from www.redditstatic.com in production,
it's breaking tinycon's ability to render a new favicon.  Instead, we'll
embed the icon directly as a data uri.  This also fixes an issue where
the main icon provides a 32x32 version as well which Chrome prefers and
makes the favicon look different after adding a bubble.

Adding the initial empty setBubble() call ensures that tinycon is aware
of the original image before onPageVisible calls reset().