Prevent timetext reflow jitter with fixed-width time column.
As the timestamps, especially the first one with its larger text, update
the time column would jump around in size.  This was very jarring and
annoying. This fixes that reflow by making the time column fixed size.

The max-width is necessary in addition to width because browsers will
auto-grow the table column without it.