Rename "reporters" to "contributors".
The more generic term allows for mixes of professional and amateur
contributors in the same event without legal confusion. It also covers
more cases since for game events it's more like "announcer" or
"commentator" etc.

The term only really shows up in the behind-the-scenes stuff
(contributor-only pages and in the code) and the sidebar now says
"updated by".