Split out a base page that isn't event-context dependent.
This pulls some of the basic liveupdate stuff out to LiveUpdatePage and
moves the event-specific page stuff into LiveUpdateEventPage.  The
latter name was already used but that was a terrible name so the old
LiveUpdateEventPage becomes LiveUpdateEventApp which is closer to the
truth anyway.

The goal here is to have liveupdate CSS and JS available in non-event
contexts for e.g. listings of events.