Only replace standalone links with embeds.
The scraper code will only process links in their own paragraph.
However, the JS would replace any link with the correct URL with the
embed. This meant that if the URL for an embedded resource were also
linked in a paragraph of explanitory text, it'd get replaced
unexpectedly with the embed.