Count activity preemptively in first request
This makes it so we record activity in the initial pageview of a live
thread in addition to the existing pixel stuff. This allows us to get a
more accurate activity count when the client does not use the pixel
endpoint, but still won't be accurate if they stay around longer than
the 15 minutes their visit will count in the activity system.