Allow caching of the about endpoint.
When a live thread is featured on reddit's homepage, we effectively DDoS
ourselves with requests to both a live thread and the thread's about page. The
about page is only ever modified when the sidebar of said thread is modified
and therefore is safe to cache for short time periods.