Read at higher consistency level in liveupdate_scraper_q
The queue is triggered immediately after new updates are posted. They
are sometimes not present on all nodes immediately. This breaks the
queue processor and makes the update never get embeds. This is in a
queue so we can afford slower queries, so let's just read at the higher
CL and make sure we get the data when we need it.