Remove _fullname override from LiveUpdateEvent
This removes an override and uses the proper upstream _fullname
implementation which does LiveUpdateEvent_{id}. This is less surprising
and more consistent.

NOTE: This is a breaking change for event listings on the site, as the
before/after parameter values will now be different, but since those are
all employee-gated right now that's fine.