Add view counts integration for live threads
This displays live total view counts in the status bar. A configurable
minimum date is used to ensure that only threads that we have full data
for have a view count shown; this prevents inaccurate or incomplete
counts from confusing users. The view count is updated once per minute
over websocket via the activity job. The view count will be visible
forever in the status bar of closed threads, and will continue to grow
if people view the thread later on.